requirements.txt
altair==5.5.0
attrs==24.2.0
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
filelock==3.16.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.26.3
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.14.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
pyarrow==18.1.0
pydeck==0.9.1
Pygments==2.18.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
safetensors==0.4.5
six==1.16.0
smmap==5.0.1
streamlit==1.40.2
tenacity==9.0.0
tokenizers==0.20.3
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
